updating save() and keep() unit tests

+++ fiftyone/core/patches.py
@@ -225,13 +225,6 @@ class _PatchesView(fov.DatasetView):
         else:
             fields = [l for l in fields if l in self._label_fields]
 
-        #
-        # IMPORTANT: we sync the contents of `_patches_dataset`, not `self`
-        # here because the `save()` call above updated the dataset, which means
-        # this view may no longer have the same contents (e.g., if `skip()` is
-        # involved)
-        #
-
         self._sync_source_root(fields)
 
     def keep(self):
@@ -244,16 +237,12 @@ class _PatchesView(fov.DatasetView):
             it immediately writes the requested changes to the underlying
             dataset.
         """
+
+        # The `keep()` operation below will delete patches, so we must sync
+        # deletions to the source dataset first
+        self._sync_source_root(self._label_fields, update=False, delete=True)
+
         super().keep()
-
-        #
-        # IMPORTANT: we sync the contents of `_patches_dataset`, not `self`
-        # here because the `save()` call above updated the dataset, which means
-        # this view may no longer have the same contents (e.g., if `skip()` is
-        # involved)
-        #
-
-        self._sync_source_root(self._label_fields)
 
     def reload(self):
         """Reloads this view from the source collection in the database.
@@ -307,42 +296,34 @@ class _PatchesView(fov.DatasetView):
 
         self._source_collection._set_labels(field, sample_ids, docs)
 
-    def _sync_source_root(self, fields):
+    def _sync_source_root(self, fields, update=True, delete=False):
         for field in fields:
-            self._sync_source_root_field(field)
+            self._sync_source_root_field(field, update=update, delete=delete)
 
-    def _sync_source_root_field(self, field):
+    def _sync_source_root_field(self, field, update=True, delete=False):
         _, label_id_path = self._get_label_field_path(field, "id")
         label_path = label_id_path.rsplit(".", 1)[0]
 
-        #
-        # Sync label updates
-        #
-
-        sample_ids, docs, label_ids = self._patches_dataset.aggregate(
-            [
-                foa.Values(self._id_field),
-                foa.Values(label_path, _raw=True),
-                foa.Values(label_id_path, unwind=True),
-            ]
-        )
-
-        self._source_collection._set_labels(field, sample_ids, docs)
-
-        #
-        # Sync label deletions
-        #
-
-        _, src_id_path = self._source_collection._get_label_field_path(
-            field, "id"
-        )
-        src_ids = self._source_collection.values(src_id_path, unwind=True)
-        delete_ids = set(src_ids) - set(label_ids)
-
-        if delete_ids:
-            self._source_collection._delete_labels(
-                ids=delete_ids, fields=field
+        if update:
+            sample_ids, docs, label_ids = self._patches_dataset.aggregate(
+                [
+                    foa.Values(self._id_field),
+                    foa.Values(label_path, _raw=True),
+                    foa.Values(label_id_path, unwind=True),
+                ]
             )
+
+            self._source_collection._set_labels(field, sample_ids, docs)
+
+        if delete:
+            all_ids = self._patches_dataset.values(label_id_path, unwind=True)
+            self_ids = self.values(label_id_path, unwind=True)
+            delete_ids = set(all_ids) - set(self_ids)
+
+            if delete_ids:
+                self._source_collection._delete_labels(
+                    ids=delete_ids, fields=field
+                )
 
 
 class PatchesView(_PatchesView):
